+++ cypress.config.ts
@@ -26,7 +26,6 @@ export default defineConfig({
       allureAttachRequests: true,
       allureCompactAttachments: 'false',
       allureResults: 'allure-results', // for test results to write
-      allureResultsWatchPath: 'allure-results/watch', // for test ops to watch
       allureSkipCommands: '', // separated comma
       // allureSkipSteps: '"after each" hook*,"before each" hook*,"before all" hook', // separated comma
       allureAddVideoOnPass: 'true',